pyproject.toml
[project]
name = "admin-project"
version = "0.0.1"
description = "The basic blog app built in the Flask tutorial."
requires-python = ">=3.8"
dependencies = [
        "alembic==1.13.2",
        "bcrypt==4.1.3",
        "blinker==1.8.2",
        "click==8.1.7",
        "colorama==0.4.6",
        "dnspython==2.6.1",
        "email-validator==2.2.0",
        "Faker==26.0.0",
        "flask==3.0.3",
        "Flask-Admin==1.6.1",
        "Flask-DebugToolbar==0.15.1",
        "Flask-Login==0.6.3",
        "Flask-Migrate==4.0.7",
        "Flask-Principal==0.4.0",
        "Flask-Security-Too==5.4.3",
        "flask-sqlalchemy==3.1.1",
        "flask-wtf==1.2.1",
        "greenlet==3.0.3",
        "idna==3.7",
        "importlib-metadata==8.0.0",
        "importlib-resources==6.4.0",
        "itsdangerous==2.2.0",
        "jinja2==3.1.4",
        "Mako==1.3.5",
        "MarkupSafe==2.1.5",
        "packaging==24.1",
        "passlib==1.7.4",
        "psycopg2-binary==2.9.9",
        "python-dateutil==2.9.0.post0",
        "python-decouple==3.8",
        "python-dotenv==1.0.1",
        "six==1.16.0",
        "SQLAlchemy==2.0.31",
        "typing-extensions==4.12.2",
        "werkzeug==3.0.3",
        "wtforms==3.1.2",
        "zipp==3.19.2",
]

# [build-system]
# requires = ["flit_core<4"]
# build-backend = "flit_core.buildapi"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"